Prior did not understand how bash_completion worked.
This file automatically sources the one inside of the home
directory. That means that we really can use this to run through all of
the files in the .bash_completion.d directory.

Additionally the Homebrew PATH needs to be figured out properly for the
optional software to be loaded up.