Completely restructure to use the `use-package' module.
The idea here is that I would like all of my packages to be loaded when
they are first initialized. But in addition to this there may be
specific configurations that are, obviously, only effective when the
package itself was loaded up.

Prior to this I felt that my custom.el was completely and utterly in
shambles. Hopefully now this will work a little bit better.